refactor(utils): extract shared httpError and getCloudApiKey helpers
