+++ utils.ts
@@ -1,3 +1,5 @@
+import type { ExtensionContext } from "@earendil-works/pi-coding-agent";
+
 export async function fetchJsonWithTimeout<T>(
   url: string,
   init: RequestInit,
@@ -74,4 +76,38 @@ export function getContextLength(modelInfo: Record<string, unknown>): number {
     }
   }
   return 128000;
+}
+
+/**
+ * Resolve the Ollama Cloud API key for a tool execution or command.
+ *
+ * Prefers the canonical provider auth chain (ctx.modelRegistry.getApiKeyForProvider),
+ * which honors runtime/CLI key overrides, the registered
+ * apiKey: "$OLLAMA_API_KEY" config, and stored auth.json credentials. Falls back
+ * to the OLLAMA_API_KEY env var for the case where the provider is not yet
+ * registered at call time.
+ */
+export async function getCloudApiKey(ctx: Pick<ExtensionContext, "modelRegistry">): Promise<string | undefined> {
+  return (await ctx.modelRegistry.getApiKeyForProvider("ollama-cloud")) ?? process.env.OLLAMA_API_KEY;
+}
+
+/**
+ * Throw a user-facing error for a non-ok Ollama Cloud HTTP response, mapping
+ * distinct status codes. Shared by the web tools and the usage command.
+ */
+export function httpError(op: string, status: number, error?: string): never {
+  if (status === 401 || status === 403) {
+    throw new Error(
+      `Ollama Cloud ${op} failed: authentication error. Check your API key in OLLAMA_API_KEY or auth.json.`,
+    );
+  }
+  if (status === 429) {
+    throw new Error(`Ollama Cloud ${op} failed: rate limited. Try again shortly.`);
+  }
+  if (status >= 500) {
+    throw new Error(`Ollama Cloud ${op} failed: server error (status ${status}). Try again shortly.`);
+  }
+  throw new Error(
+    `Ollama Cloud ${op} failed: unexpected response (status ${status}${error ? `: ${error}` : ""}). Try again shortly.`,
+  );
 }
